+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - **add:** error message file
 - **fix:** adjust test setup
 - **add:** tests for the general usage of the reply interface
+- **fix:** adjust test setup
 
 ### [v0.0.2] - 2016-06-24
 - **doc:** change logo title